requirements.txt
langchain
langchain-core
langchain-community
langchain-chroma
langchain-huggingface
langchain-groq
chromadb
pypdf
python-dotenv
streamlit

